Refactor(auth): Move and refactor `AuthService` to `services` directory
